+++ src/check/rules/term/meta.rs
@@ -1,7 +1,11 @@
-use crate::check::monad::{TCE, TCMS, TCS};
-use crate::syntax::core::{Elim, Term, Val};
+use voile_util::meta::{MetaSolution, MI};
 
-pub trait HasMeta {
+use crate::check::monad::{TCE, TCMS, TCS};
+use crate::check::rules::term::simplify;
+use crate::syntax::common::Bind;
+use crate::syntax::core::{Closure, Elim, Term, Val};
+
+pub trait HasMeta: Sized {
     /// Inline solved metas inside `self`.
     fn inline_meta(self, tcs: TCS) -> TCMS<Self>;
 }
@@ -30,6 +34,8 @@ impl HasMeta for Elim {
 impl HasMeta for Term {
     fn inline_meta(self, tcs: TCS) -> TCMS<Self> {
         match self {
+            // Prefer not to simplify
+            Term::Whnf(Val::Meta(mi, elims)) => solve_meta(tcs, mi, elims),
             Term::Whnf(w) => w.inline_meta(tcs).map(|(w, tcs)| (Term::Whnf(w), tcs)),
             Term::Redex(gi, id, elims) => {
                 let (elims, tcs) = elims.inline_meta(tcs)?;
@@ -39,15 +45,50 @@ impl HasMeta for Term {
     }
 }
 
+impl<T: HasMeta> HasMeta for Bind<T> {
+    fn inline_meta(self, tcs: TCS) -> TCMS<Self> {
+        let (ty, tcs) = self.ty.inline_meta(tcs)?;
+        Ok((Bind::new(self.licit, self.name, ty), tcs))
+    }
+}
+
+impl HasMeta for Closure {
+    fn inline_meta(self, tcs: TCS) -> TCMS<Self> {
+        match self {
+            Closure::Plain(body) => body
+                .inline_meta(tcs)
+                .map(|(b, tcs)| (Closure::plain(b), tcs)),
+        }
+    }
+}
+
+fn solve_meta(tcs: TCS, mi: MI, elims: Vec<Elim>) -> TCMS<Term> {
+    use MetaSolution::*;
+    let sol = match tcs.meta_context.solution(mi) {
+        Solved(sol) => sol.clone(),
+        Unsolved => Err(TCE::MetaUnsolved(mi))?,
+        Inlined => unreachable!(),
+    };
+    let (elims, tcs) = elims.inline_meta(tcs)?;
+    Ok((sol.apply_elim(elims), tcs))
+}
+
 impl HasMeta for Val {
     fn inline_meta(self, tcs: TCS) -> TCMS<Self> {
         use Val::*;
         match self {
             Type(l) => Ok((Type(l), tcs)),
             Data(k, gi, args) => args.inline_meta(tcs).map(|(a, tcs)| (Data(k, gi, a), tcs)),
-            Pi(_, _) => unimplemented!(),
+            Pi(t, clos) => {
+                let (t, tcs) = t.unboxed().inline_meta(tcs)?;
+                let (clos, tcs) = clos.inline_meta(tcs)?;
+                Ok((Val::Pi(t.boxed(), clos), tcs))
+            }
             Cons(c, ts) => ts.inline_meta(tcs).map(|(ts, tcs)| (Cons(c, ts), tcs)),
-            Meta(_, _) => unimplemented!(),
+            Meta(mi, elims) => {
+                let (sol, tcs) = solve_meta(tcs, mi, elims)?;
+                simplify(tcs, sol)
+            }
             Axiom(a) => Ok((Axiom(a), tcs)),
             Var(head, args) => args.inline_meta(tcs).map(|(a, tcs)| (Var(head, a), tcs)),
             Id(t, a, b) => {