+++ src/check/rules/term/mod.rs
@@ -6,6 +6,7 @@ use crate::syntax::abs::Abs;
 use crate::syntax::core::{Bind, Closure, Term, Val};
 
 pub use self::infer::infer;
+pub use self::meta::HasMeta;
 pub use self::unify::subtype;
 pub use self::whnf::simplify;