+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # 0.0.6
 
++ Fix a bunch of index-relevant problems
+
 # 0.0.5
 
 + Migrate from CircleCI + AppVeyor to GitHub Actions