+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 + Pretty print `Abs`
 + Support tracing the type-checking process
 + Put `infer`, `unify` and `whnf` under `check::rules::term`
++ Inline meta variables after its declaration is checked
 
 # 0.0.5